added more log messages to update_return
