+++ cwe_checker_rs/src/analysis/graph.rs
@@ -50,21 +50,26 @@ pub type Graph<'a> = DiGraph<Node<'a>, Edge<'a>>;
 /// The node type of an interprocedural control flow graph
 ///
 /// Each node carries a pointer to its associated block with it.
-/// For `CallReturn`nodes the associated block is the callsite block (containing the call instruction)
-/// and *not* the return block (containing the return instruction).
+/// For `CallReturn`nodes the associated blocks are both the callsite block (containing the call instruction)
+/// and the returning-from block (containing the return instruction).
 #[derive(Serialize, Debug, PartialEq, Eq, Hash, Clone, Copy)]
 pub enum Node<'a> {
     BlkStart(&'a Term<Blk>),
     BlkEnd(&'a Term<Blk>),
-    CallReturn(&'a Term<Blk>),
+    CallReturn {
+        call: &'a Term<Blk>,
+        return_: &'a Term<Blk>,
+    },
 }
 
 impl<'a> Node<'a> {
-    /// Get the block corresponding to the node.
+    /// Get the block corresponding to the node for `BlkStart` and `BlkEnd` nodes.
+    /// panics if called on a `CallReturn` node.
     pub fn get_block(&self) -> &'a Term<Blk> {
         use Node::*;
         match self {
-            BlkStart(blk) | BlkEnd(blk) | CallReturn(blk) => blk,
+            BlkStart(blk) | BlkEnd(blk) => blk,
+            CallReturn { .. } => panic!("get_block() is undefined for CallReturn nodes"),
         }
     }
 }
@@ -74,7 +79,11 @@ impl<'a> std::fmt::Display for Node<'a> {
         match self {
             Self::BlkStart(block) => write!(formatter, "BlkStart @ {}", block.tid),
             Self::BlkEnd(block) => write!(formatter, "BlkEnd @ {}", block.tid),
-            Self::CallReturn(block) => write!(formatter, "CallReturn (caller @ {})", block.tid),
+            Self::CallReturn { call, return_ } => write!(
+                formatter,
+                "CallReturn @ {} (caller @ {})",
+                return_.tid, call.tid
+            ),
         }
     }
 }
@@ -203,8 +212,7 @@ impl<'a> GraphBuilder<'a> {
 
     /// Add all outgoing edges generated by calls and interprocedural jumps for a specific block to the graph.
     /// Return edges are *not* added by this function.
-    fn add_outgoing_edges(&mut self, node: NodeIndex) {
-        let block: &'a Term<Blk> = self.graph[node].get_block();
+    fn add_outgoing_edges(&mut self, node: NodeIndex, block: &'a Term<Blk>) {
         let jumps = block.term.jmps.as_slice();
         match jumps {
             [] => (), // Blocks without jumps are dead ends corresponding to control flow reconstruction errors.
@@ -231,13 +239,17 @@ impl<'a> GraphBuilder<'a> {
         }
         for (call_node, return_to_node) in self.return_addresses[&return_from_sub.tid].iter() {
             let call_block = self.graph[*call_node].get_block();
+            let return_from_block = self.graph[return_source].get_block();
             let call_term = call_block
                 .term
                 .jmps
                 .iter()
                 .find(|jump| matches!(jump.term.kind, JmpKind::Call(_)))
                 .unwrap();
-            let cr_combine_node = self.graph.add_node(Node::CallReturn(call_block));
+            let cr_combine_node = self.graph.add_node(Node::CallReturn {
+                call: call_block,
+                return_: return_from_block,
+            });
             self.graph
                 .add_edge(*call_node, cr_combine_node, Edge::CRCallStub);
             self.graph
@@ -267,8 +279,8 @@ impl<'a> GraphBuilder<'a> {
     /// Add all non-return-instruction-related jump edges to the graph.
     fn add_jump_and_call_edges(&mut self) {
         for node in self.graph.node_indices() {
-            if let Node::BlkEnd(_) = self.graph[node] {
-                self.add_outgoing_edges(node);
+            if let Node::BlkEnd(block) = self.graph[node] {
+                self.add_outgoing_edges(node, block);
             }
         }
     }
@@ -298,8 +310,8 @@ pub fn get_indices_of_block_nodes<'a, I: Iterator<Item = &'a Tid>>(
     let tids: HashSet<Tid> = block_tids.cloned().collect();
     let mut tid_to_indices_map = HashMap::new();
     for node_index in graph.node_indices() {
-        if let Some(tid) = tids.get(&graph[node_index].get_block().tid) {
-            if let Node::BlkStart(_block_term) = graph[node_index] {
+        if let Node::BlkStart(block_term) = graph[node_index] {
+            if let Some(tid) = tids.get(&block_term.tid) {
                 let start_index = node_index;
                 let end_index = graph.neighbors(start_index).next().unwrap();
                 tid_to_indices_map.insert(tid.clone(), (start_index, end_index));
